Update card.go
Added some comments for clarity as I was learning the syntax.
Does the last comment take and return a string? Seems like it should be taking the facevalue from the object. Idk if go has the option for like "this' like java does
so that it would just take it's own facevalue instead of having to pipe it in